--- src/components/Crumb/Crumb.module.scss
@@ -1,25 +1,0 @@
-@use '../../styles/core/boilerplate.scss'as *;
-
-.top_right {
-  position: fixed;
-  top: 5px;
-  right: 5px;
-}
-
-.top_left {
-  position: fixed;
-  top: 5px;
-  left: 5px;
-}
-
-.bottom_right {
-  position: fixed;
-  bottom: 5px;
-  right: 5px;
-}
-
-.bottom_left {
-  position: fixed;
-  bottom: 5px;
-  left: 5px;
-}
