--- src/components/Crumb/icons/icon.module.scss
@@ -1,21 +1,0 @@
-@use '../../../styles/core/boilerplate.scss'as *;
-
-.success,
-.error,
-.info {
-  fill: $color-text-reverse;
-}
-
-.warning {
-  fill: $color-text-main;
-}
-
-.custom-success,
-.custom-error,
-.custom-info {
-  color: $color-text-reverse;
-}
-
-.custom-warning {
-  color: $color-text-main;
-}
