--- src/components/Crumb/icons/index.ts
@@ -1,5 +1,0 @@
-export { default as InfoIcon } from './InfoIcon';
-export { default as SuccessIcon } from './SuccessIcon';
-export { default as WarningIcon } from './WarningIcon';
-export { default as ErrorIcon } from './ErrorIcon';
-export { default as CustomIcon } from './CustomIcon';
