--- src/components/Crumb/index.ts
@@ -1,7 +1,0 @@
-export {
-  default,
-  CrumbInterfaceT,
-  NewCrumbT,
-  CrumbTypesE,
-  CrumbLocationE,
-} from './Crumb';
