+++ src/components/Input/Input.module.scss
@@ -73,7 +73,7 @@
 
 @mixin icon() {
   position: absolute;
-  top: 42px;
+  top: 44px;
   right: 12px;
 }
