+++ src/components/Notification/types.ts
@@ -1,0 +1,14 @@
+export enum NotificationTypesE {
+  NEUTRAL = 'neutral',
+  SUCCESS = 'success',
+  ERROR = 'error',
+  INFO = 'info',
+  WARNING = 'warning',
+}
+
+export enum NotificationLocationE {
+  TOP_RIGHT = 'top_right',
+  TOP_LEFT = 'top_left',
+  BOTTOM_RIGHT = 'bottom_right',
+  BOTTOM_LEFT = 'bottom_left',
+}
