--- src/components/Toast/ToastMessage.module.scss
@@ -1,89 +1,0 @@
-@use '../../styles/core/boilerplate.scss'as *;
-
-.wrapper {
-  position: absolute;
-  top: 5px;
-  right: 5px;
-}
-
-.container {
-  z-index: 100;
-  padding: 16px;
-  width: 523px;
-  position: relative;
-  overflow: hidden;
-  box-shadow: $drop-shadow;
-  margin-bottom: 14px;
-  background-color: $color-background-neutral-0;
-  color: $color-text-main;
-  box-sizing: border-box;
-
-  @include mobile-down {
-    margin: 0;
-    width: 100vw;
-    max-width: 100vw;
-  }
-
-  &.success {
-    .icon {
-      stroke: $color-semantic-success;
-    }
-    .progress_bar {
-      background-color: $color-semantic-success;
-    }
-  }
-  &.error {
-    .icon {
-      stroke: $color-semantic-critical;
-    }
-    .progress_bar {
-      background-color: $color-semantic-critical;
-    }
-  }
-  &.warning {
-    .icon {
-      stroke: $color-semantic-warning;
-    }
-    .progress_bar {
-      background-color: $color-semantic-warning;
-    }
-  }
-  &.info {
-    .icon {
-      stroke: $color-semantic-info;
-    }
-    .progress_bar {
-      background-color: $color-semantic-info;
-    }
-  }
-}
-
-.close {
-  position: absolute;
-  right: 5px;
-  top: 5px;
-  color: $color-text-main;
-}
-
-.progress_bar {
-  position: absolute;
-  height: 8px;
-  width: calc(100% * var(--progress));
-  bottom: 0;
-  left: 0;
-  right: 0;
-
-  &.active {
-    border-top-right-radius: 5px;
-    border-bottom-right-radius: 5px;
-  }
-}
-
-.callback_wrapper {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.icon {
-  margin-right: 20px;
-}
