--- src/components/Toast/icons/index.ts
@@ -1,4 +1,0 @@
-export { default as InfoIcon } from './InfoIcon';
-export { default as SuccessIcon } from './SuccessIcon';
-export { default as WarningIcon } from './WarningIcon';
-export { default as ErrorIcon } from './ErrorIcon';
