--- src/components/Toast/index.ts
@@ -1,7 +1,0 @@
-export {
-  default,
-  ToastInterfaceT,
-  NewToastT,
-  ToastTypesE,
-  ToastLocationE,
-} from './Toast';
