+++ src/components/index.ts
@@ -1,55 +1,39 @@
 export { default as Avatar } from './Avatar';
 export { default as Badge, BadgeCategoriesE } from './Badge';
-export { default as Dropdown, AlignmentT, dropdownT } from './Dropdown';
+export { default as Dropdown, type AlignmentT, type dropdownT } from './Dropdown';
 export {
   default as Button,
-  ButtonT,
+  type ButtonT,
   ButtonCategoriesE,
   ButtonSizesE,
 } from './Button';
 export { default as Checkbox } from './Checkbox';
-
 export { default as CopyButton } from './CopyButton';
 export {
-  default as Crumb,
-  CrumbInterfaceT,
-  NewCrumbT,
-  CrumbTypesE,
-  CrumbLocationE,
-} from './Crumb';
-export {
   default as Notification,
-  NotificationInterfaceT,
-  NewNotificationT,
+  type NotificationInterfaceT,
+  type NewNotificationT,
   NotificationTypesE,
   NotificationLocationE,
 } from './Notification';
 export { default as Icon, IconT } from './Icon';
 export {
   default as Input,
-  InputInterfaceT,
-  InputT,
+  type InputInterfaceT,
+  type InputT,
   InputIconColorE,
 } from './Input';
 export { default as ProgressBar } from './ProgressBar';
 export { default as RadioButton } from './RadioButton';
-
-export { default as Select, OptionT, SelectInterfaceT } from './Select';
+export { default as Select, type OptionT, type SelectInterfaceT } from './Select';
 export { default as Switch } from './Switch';
 export {
   default as TextArea,
-  TextAreaInterfaceT,
+  type TextAreaInterfaceT,
   TextAreaIconColorE,
 } from './TextArea';
 export {
-  default as Toast,
-  ToastInterfaceT,
-  NewToastT,
-  ToastTypesE,
-  ToastLocationE,
-} from './Toast';
-export {
   default as ToolTip,
   ToolTipLocationE,
-  ToolTipCategoriesE,
+  ToolTipCategoriesE, 
 } from './ToolTip';
